Fix video tooltip (#4916)
* move modal looker state to use-looker

* fix video update sample

* linting
